--- camera/d435/README.md
@@ -1,1 +1,0 @@
-# D435 Runner
