+++ scripts/d435_record_rosbag_user.py
@@ -8,17 +8,8 @@
 
 """
 
-import cv2
-import numpy as np
-import os
-import argparse
-
-import camera.d435.d435_runner as d435
-import camera.utils.display as display
-from camera.utils.writer_ros import vidWriter_ROS
-
-from d435_recorders import D435RecRosbag
-
+import camera.d435.runner as d435
+from .d435_recorders import D435RecRosbag
 
 
 # prepare
