+++ scripts/vid_compress.py
