+++ scripts/vid_compress_bag.py
